backend/requirements.txt
# Core API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AWS SDK
boto3==1.34.34
botocore==1.34.34

# Audio Processing
numpy==1.26.3
librosa==0.10.1
soundfile==0.12.1
pyloudnorm==0.1.1
noisereduce==3.0.0
scipy==1.12.0

# ML/AI
torch==2.1.2
torchaudio==2.1.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
redis==5.0.1

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
